+++ RELEASE_NOTES.md
@@ -61,6 +61,7 @@ The required Doxygen version for documentation generation is version 1.11.
 - (wifi) Fix S-MPDU TX duration computation with BlockAck ack policy
 - (wifi) Fix missing DSSS Param Set in Probe Request sent over 2.4 GHz links
 - (spectrum) Trigger 3GPP channel updates after changes to antenna attributes
+- (wifi) #1104 - Fix wrong calculation of start sequence number
 
 ## Release 3.43